把xxd打印出来的amiibo的540大小的数组成功存入W25Q64FV中,并且将W25Q64FV的写入函数修改成每写满一页就跳转到下一页.

+++ Amiibo/User/hal/spi_port.h
@@ -8,9 +8,9 @@
 #define LSB 0
 
 typedef struct {
-    uint8_t len;
+    uint32_t len;
     uint8_t flags;
-    uint8_t get_buf_len;
+    uint32_t get_buf_len;
     uint8_t bit_oder;
     uint8_t *buf;
     uint8_t *getbuf;
